Fundamental Bug Fix
Found and fixed two bugs:
Repeat processing of emails --> When receiving multiple Calendly emails with the same subject, all emails--regardless of whether they're in trash or not--are considered on the same thread.
Previously, multi_cal did not mark emails as "read" when it deleted them, this was causing some already-processed emails to be re-read and re-evaluated erroneously (causing phantom bookings and cancellations)

Group event occupancy tags were also being double-cancelled when the group event was fully cancelled--once by the main adjustSchedule function and once by the adjustGroupEvent function. The adjustGroupEvent portion has been removed.